Move reimbursements data minimisation to apps
